pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "stixdb-engine"
version = "0.1.0"
description = "StixDB — Reasoning Agentic Context Database. An autonomous AI agent lives inside the DB itself, managing a self-organizing graph memory layer for other agents."
authors = [{ name = "StixDB" }]
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["ai", "agent", "database", "graph", "memory", "llm", "rag"]

dependencies = [
    # Core graph
    "networkx>=3.2",
    # Ingestion
    "pypdf>=4.2",
    # LLM providers (all optional at runtime)
    "openai>=1.30",
    "anthropic>=0.25",
    "httpx>=0.27",
    # API server
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "pydantic>=2.7",
    # Async scheduling
    "apscheduler>=3.10",
    # Observability
    "prometheus-client>=0.20",
    # Utilities
    "numpy>=1.26",
    "tenacity>=8.3",
    "structlog>=24.1",
    "python-dotenv>=1.0",
    "typer>=0.12",
    "rich>=13.7",
    "python-multipart>=0.0.9",
    "minio>=7.2"
]

[project.optional-dependencies]
qdrant  = ["qdrant-client>=1.9"]
local   = ["sentence-transformers>=2.6", "chromadb>=0.5"]
kuzu    = ["kuzu>=0.6"]
neo4j   = ["neo4j>=5.19"]
sql     = ["sqlalchemy[asyncio]>=2.0", "psycopg2-binary>=2.9"]
# Convenience group for local dev (KuzuDB + sentence-transformers, no Docker)
local-dev = ["kuzu>=0.6", "sentence-transformers>=2.6"]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "mypy>=1.10",
    "ruff>=0.4",
]

[project.scripts]
stixdb = "stixdb.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["stixdb"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
